Refactored NativeScriptComponent
