+++ OloEngine/src/CMakeLists.txt
@@ -78,6 +78,7 @@
 		"OloEngine/Scene/SceneSerializer.cpp"
 		"OloEngine/Scene/SceneSerializer.h"
 		"OloEngine/Scene/NativeScript.h"
+		"OloEngine/Scene/NativeScriptComponent.h"
 
 		"OloEngine/Utils/PlatformUtils.h"
