+++ OloEngine/src/OloEngine/Scene/Scene.cpp
@@ -4,6 +4,7 @@
 #include "Scene.h"
 
 #include "Components.h"
+#include "NativeScriptComponent.h"
 #include "OloEngine/Renderer/Renderer2D.h"
 
 #include <glm/glm.hpp>
